AI_LLM/RAG/RAG-Cancer/requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
chromadb
pypdf
python-dotenv
sentence-transformers
